Rename `tactic/stop` folder to `tactic/halt`, renaming addresses where necessary
