+++ src/software/ai/hl/stp/play/enemy_free_kick/BUILD
@@ -26,7 +26,7 @@ cc_library(
         "//software/ai/hl/stp/tactic/goalie:goalie_tactic",
         "//software/ai/hl/stp/tactic/move:move_tactic",
         "//software/ai/hl/stp/tactic/pass_defender:pass_defender_tactic",
-        "//software/ai/hl/stp/tactic/stop:halt_tactic",
+        "//software/ai/hl/stp/tactic/halt:halt_tactic",
         "//software/logger",
         "//software/util/generic_factory",
         "//software/world:game_state",
