+++ src/software/ai/hl/stp/tactic/BUILD
@@ -28,7 +28,7 @@ cc_library(
         "//software/ai/hl/stp/tactic/pivot_kick:pivot_kick_tactic",
         "//software/ai/hl/stp/tactic/receiver:receiver_tactic",
         "//software/ai/hl/stp/tactic/shadow_enemy:shadow_enemy_tactic",
-        "//software/ai/hl/stp/tactic/stop:halt_tactic",
+        "//software/ai/hl/stp/tactic/halt:halt_tactic",
     ],
 )
