+++ src/software/ai/hl/stp/tactic/halt/BUILD
